Rename controllers to *.controller.ts
